Move `Os` class to utils package
The `common` package is public and this class is
internal.
